Reduce required coverage
While coverage is a good helping tool to check the total state of testing in your project, requiring 100% is not very helpful as it enforces tests for parts that are very brittle to test (e.g. for pure wiring code). These tests can even harm projects. More interesting and helpful than the coverage numbers itself are the trends of these numbers (going up, going down)