Replace function calls in tests by arrow functions
As noted in an earlier commit, the functions were in place as mocha tries to bind some context to it, that enables some of mochas features. This does not work the same way with arrow functions. As these features from mocha are most of the time not required (and sometimes even add arguable sideeffect behavior) these function () callbacks were replaced in this commit